Separated examples, predictions and classifications into a bit more logical structures, but there is still work to be done here. Updated CorpusVisualizer to use the new classification-structures. Added command-line classifier parameters to SplitAnalysis.py.
Also wrote a new exampleBuilder (SimpleDependencyExampleBuilder2.py). Unlike the previous version, this class adds to the features the dependency type, and tags the terminus tokens text and POS according to the dependency's direction. In other words, dependency directions are now implemented. The F-score for the classification is over 80 %, there are either bugs at work here, or this stuff really works.

Also improved the output options. Now the program outputs all kinds of intermediate and debug data to the defined folder.