First version, seems to work. The approach is generally this: Go through all predicted things, look for matching ones in gold standard. Matched predictions are true positives, unmatched predictions are false positives. All of the things in gold standard that didn't get matched are false negatives.
This has the consequence that when there is e.g. a predicted entity of type A and a gold entity with type B that have the same head offsets, in previous approaches this would simply be one false positive example. However, here this results in one false positive from the A prediction and one false negative from the unmatched gold standard B. This means that at least micro-averages might differ slightly from previous evaluation approaches.