Fixed a bug related to example class determination. The code now correctly gives a negative class to all edges connecting to dummy-entities.
Previously, it gave the negative class to everything but edges connecting to dummy-entities. Thus the remaining examples should have been negatives, but since the early-out rule of "check for a dummy" worked in reverse,  the dummy-entities got used in defining the type of the example, causing the system to effectively produce in triplicate the examples that the normal edge example builder would produce from an unflattened input file.

Commented out unused levelByInteraction dictionary.